requirements.txt
python-dotenv==0.19.2
aiogram==2.19
psycopg2==2.9.3